apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
rq==1.15.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
faker==20.1.0